Audio helper

+++ src/Protocol.cpp
@@ -101,7 +101,7 @@ String Protocol::buildDebugResponseMessage()
     return buffer;
 }
 
-String Protocol::buildKothConfigClient(uint8_t maxPoints, uint8_t countdown, uint8_t captureTime, uint8_t maxTime, uint8_t respawnTime)
+String Protocol::buildKothConfigClient(uint16_t maxPoints, uint16_t countdown, uint16_t captureTime, uint16_t maxTime, uint16_t respawnTime)
 {
     String buffer;
     buffer = String(CONF) + ";";
@@ -114,7 +114,7 @@ String Protocol::buildKothConfigClient(uint8_t maxPoints, uint8_t countdown, uin
     return buffer;
 }
 
-String Protocol::buildKothConfig(uint8_t maxPoints, uint8_t countdown, uint8_t captureTime, uint8_t maxTime,uint8_t respawnTime)
+String Protocol::buildKothConfig(uint16_t maxPoints, uint16_t countdown, uint16_t captureTime, uint16_t maxTime,uint16_t respawnTime)
 {
     String buffer;
     buffer = String(CONF) + ";";
